extract GitCheckout type to a testing module
Summary:
This will be used to set up the sapling tests.

Test Plan: Rely on CI.

Reviewers: suo

Subscribers:

Tasks:

Tags:
