+++ src/git.rs
@@ -175,106 +175,8 @@ pub fn get_paths_from_cmd(paths_cmd: &str) -> Result<Vec<AbsPath>> {
 
 #[cfg(test)]
 mod tests {
-    use std::{fs::OpenOptions, io::Write};
-
     use super::*;
-    use tempfile::TempDir;
-
-    struct GitCheckout {
-        root: TempDir,
-    }
-
-    impl GitCheckout {
-        fn new() -> Result<GitCheckout> {
-            let git = GitCheckout {
-                root: TempDir::new()?,
-            };
-
-            assert_eq!(git.run("init").status()?.code(), Some(0));
-
-            // We add an initial commit because git diff-tree behaves
-            // differently when HEAD is the only commit in the
-            // repository. In actual production uses, our git
-            // diff-tree invocation will show the files modified in
-            // the HEAD commit compared to HEAD~, but if HEAD~ doesn't
-            // exist, it returns an empty list of files.
-            git.write_file("README", "or don't")?;
-            git.add("README")?;
-            git.commit("initial commit")?;
-
-            Ok(git)
-        }
-
-        fn rm_file(&self, name: &str) -> Result<()> {
-            let path = self.root.path().join(name);
-            std::fs::remove_file(path)?;
-            Ok(())
-        }
-
-        fn write_file(&self, name: &str, contents: &str) -> Result<()> {
-            let path = self.root.path().join(name);
-            let mut file = OpenOptions::new()
-                .read(true)
-                .append(true)
-                .create(true)
-                .open(path)?;
-
-            writeln!(file, "{}", contents)?;
-            Ok(())
-        }
-
-        fn checkout_new_branch(&self, branch_name: &str) -> Result<()> {
-            let output = Command::new("git")
-                .args(&["checkout", "-b", branch_name])
-                .current_dir(self.root.path())
-                .output()?;
-            assert!(output.status.success());
-            Ok(())
-        }
-
-        fn add(&self, pathspec: &str) -> Result<()> {
-            let output = Command::new("git")
-                .args(&["add", pathspec])
-                .current_dir(self.root.path())
-                .output()?;
-            assert!(output.status.success());
-            Ok(())
-        }
-
-        fn commit(&self, message: &str) -> Result<()> {
-            let output = Command::new("git")
-                .args(&["commit", "-m", message])
-                .current_dir(self.root.path())
-                .output()?;
-            assert!(output.status.success());
-            Ok(())
-        }
-
-        fn changed_files(&self, relative_to: Option<&str>) -> Result<Vec<String>> {
-            std::env::set_current_dir(self.root.path())?;
-            let repo = Repo::new()?;
-            let files = repo.get_changed_files(relative_to)?;
-            let files = files
-                .into_iter()
-                .map(|abs_path| abs_path.file_name().unwrap().to_string_lossy().to_string())
-                .collect::<Vec<_>>();
-            Ok(files)
-        }
-
-        fn merge_base_with(&self, merge_base_with: &str) -> Result<String> {
-            std::env::set_current_dir(self.root.path())?;
-            let repo = Repo::new()?;
-            repo.get_merge_base_with(merge_base_with)
-        }
-
-        // Returns a Command to run the subcommand in the clone.
-        fn run(&self, subcommand: &str) -> std::process::Command {
-            let mut cmd = std::process::Command::new("git");
-            cmd.arg(subcommand);
-            cmd.current_dir(self.root.path());
-            cmd
-        }
-    }
+    use crate::testing::GitCheckout;
 
     // Should properly detect changes in the commit (and not check other files)
     #[test]
@@ -344,10 +246,7 @@ mod tests {
         git.add(".")?;
         git.commit("commit 2")?;
 
-        let output = Command::new("git")
-            .args(&["mv", "test_2.txt", "new.txt"])
-            .current_dir(git.root.path())
-            .output()?;
+        let output = git.run("mv").arg("test_2.txt").arg("new.txt").output()?;
         assert!(output.status.success());
 
         let files = git.changed_files(None)?;
